+++ src/euroeval/dataset_configs/faroese.py
@@ -2,7 +2,7 @@
 
 from ..data_models import DatasetConfig
 from ..languages import FAROESE
-from ..tasks import LA, NER, RC, SENT
+from ..tasks import GED, LA, NER, RC, SENT
 
 # Official datasets ###
 
@@ -56,6 +56,15 @@ MULTI_WIKI_QA_FO_CONFIG = DatasetConfig(
     pretty_name="MultiWikiQA-fo",
     source="EuroEval/multi-wiki-qa-fo-mini",
     task=RC,
+    languages=[FAROESE],
+    unofficial=True,
+)
+
+GERLANGMOD_FO_CONFIG = DatasetConfig(
+    name="gerlangmod-fo",
+    pretty_name="GerLangMod-fo",
+    source="EuroEval/gerlangmod-fo",
+    task=GED,
     languages=[FAROESE],
     unofficial=True,
 )