+++ src/euroeval/dataset_configs/norwegian.py
@@ -5,6 +5,7 @@ from ..languages import NORWEGIAN, NORWEGIAN_BOKMÅL, NORWEGIAN_NYNORSK
 from ..tasks import (
     COMMON_SENSE,
     EUROPEAN_VALUES,
+    GED,
     KNOW,
     LA,
     MCRC,
@@ -274,4 +275,22 @@ NORSUMM_NN_CONFIG = DatasetConfig(
     languages=[NORWEGIAN_NYNORSK, NORWEGIAN],
     unofficial=True,
     val_split=None,
+)
+
+GERLANGMOD_NB_CONFIG = DatasetConfig(
+    name="gerlangmod-nb",
+    pretty_name="GerLangMod-nb",
+    source="EuroEval/gerlangmod-nb",
+    task=GED,
+    languages=[NORWEGIAN_BOKMÅL, NORWEGIAN],
+    unofficial=True,
+)
+
+GERLANGMOD_NN_CONFIG = DatasetConfig(
+    name="gerlangmod-nn",
+    pretty_name="GerLangMod-nn",
+    source="EuroEval/gerlangmod-nn",
+    task=GED,
+    languages=[NORWEGIAN_NYNORSK, NORWEGIAN],
+    unofficial=True,
 )