+++ src/euroeval/prompt_templates/__init__.py
@@ -5,6 +5,7 @@ import typing as t
 from ..data_models import PromptConfig
 from ..languages import get_all_languages
 from .classification import CLASSIFICATION_TEMPLATES
+from .grammatical_error_detection import GED_TEMPLATES
 from .linguistic_acceptability import LA_TEMPLATES
 from .multiple_choice import MULTIPLE_CHOICE_TEMPLATES
 from .named_entity_recognition import NER_TEMPLATES